src/requirements.txt
fastapi
uvicorn
pydantic
streamlit
transformers
sentence-transformers
faiss-cpu
torch
numpy
requests
python-multipart
audio-recorder-streamlit
gtts
